⏳ add loader to submit button & refactor
